Remove asserts in MathUtils::round_int() and MathUtils::truncate_int().
These are very low level functions and get used a *lot*. As a result, even
the tests for the asserts have a measurable impact on performance, Surely
these are best left enabled only in some form of debug build? (Also, the
reaction in the case of overflow or underflow, to terminate the application,
seems severe to use in release builds to me.)

This patch simply wraps them in #if 0. I open the floor to suggestions of a
better switch.

I tested on a Raspberry Pi with the default theme, displaying the front page
with the RSS ticker enabled. This saturates the CPU, so I'm measuring the
improvement using the debug window's FPS figure. This patch improves this from
~49.8 FPS to ~50.8 FPS.